Move info and destroy under peerings
