+++ packages/cli/src/commands/spaces/peerings/info.ts
