+++ packages/cli/test/unit/commands/spaces/peerings/destroy.unit.test.ts
@@ -1,5 +1,5 @@
 import {stderr} from 'stdout-stderr'
-import Cmd from '../../../../../src/commands/spaces/peering/destroy'
+import Cmd from '../../../../../src/commands/spaces/peerings/destroy'
 import runCommand from '../../../../helpers/runCommand'
 import * as nock from 'nock'
 import heredoc from 'tsheredoc'
