+++ packages/cli/test/unit/commands/spaces/peerings/info.unit.test.ts
@@ -1,5 +1,5 @@
 import {stdout} from 'stdout-stderr'
-import Cmd from '../../../../../src/commands/spaces/peering/info'
+import Cmd from '../../../../../src/commands/spaces/peerings/info'
 import runCommand from '../../../../helpers/runCommand'
 import * as nock from 'nock'
 import heredoc from 'tsheredoc'
